backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0

# CORS para conectar React
django-cors-headers>=4.3.0

# Variables de entorno
python-decouple>=3.8

# Base de datos PostgreSQL
psycopg2-binary>=2.9.9

# Manejo de imágenes
Pillow>=10.0.0

# Autenticación con JWT
djangorestframework-simplejwt>=5.3.0
djangorestframework-simplejwt[token_blacklist]>=5.3.0

# Servidor para producción
gunicorn>=21.2.0

# Utilidades
django-filter>=24.0
pytest>=8.0.0
pytest-django>=4.8.0

# Necesario para usar postgres
psycopg2-binary

# Para que Django pueda servir sus propios archivos estaticos
whitenoise

# Cargar variables de entorno desde archivos .env
python-dotenv>=1.0.0

# Configuración de base de datos desde URL
dj-database-url>=2.1.0

# Driver PostgreSQL moderno (versión 3)
psycopg[binary]>=3.1.0
